Remove assert that might needlessly trigger
This usb implementation consists of a worker thread that executes all usb commands
by proxy for the main js loop

There is a queue to this worker, and a queue from this worker.

The queue to the worker wakes up the worker thread using a condition variable.

However, the queue FROM the worker needs something to inject it's handler into
the main event loop. This is solved using a http://docs.libuv.org/en/v1.x/async.html

A JS-program, when no timers or events are pending in the main loop, quits. However,
whenever a uv_async_t exists, the main loop cannot quit. This was a bug with the queue
from the queue, where first the uv_async_t always existed, even there were no pending
tasks being executed on the worker.

This again was solved by some reference counting. Whenever ref got > 0, we created
the uv_async. Ref got incremented whenever we put something into the queue to the worker.

Whenever we processed something from the queue from the worker, we decremented ref, and
destroyed the uv_async_t whenever ref == 0.

Sounds fine, right? NO! The handler for uv_async_t running in the js main loop tried to
process as many events available in the queue for each async event, while the worker sent
a separate signal for each item. This is a correct implementation, as each async_send can
be merged into the same wakeup.

However, it seemed to create a situation where:
1. The worker posts an answer to the from worker queue
2. The worker does an async_send to wake up
3. The main js loop picks up the event and pops the answer from the queue and starts processing,
4. Before the processing is done, the worker posts another answer to the queue
5. The main js loop is done processing the first item, and in it's while(items_still_on_quue) picks up the second answer, and this is processed
6. The main js loop is now done processing, and decrements ref by 2. Ref is now 0, and the uv_async is destroyed
7. The worker thread tries do do async_send on this nonexistent uv_async. BANG!

This patch accepts that the async can be gone in step 7.