Add retry block when getting LIBUSB_ERROR_ACCESS
When upgrading via clijs on a Windows machine with a Huddly driver
installed there is a race condition. After the first half of the upgrade
the device reboots. Upon reboot both clijs and the Windows driver access
the USB interface. clijs can reach a state where it is unable to get the
device's serial, thus failing the upgrade. In order to fix this - a
retry should be attempted when getting LIBUSB_ERROR_ACCESS.