Allow listDevices to update list when serial is 'unknown'
When listing devices the function _listDevices checks if the new list
(newList) has devices already listed in previousList. The problem is
that an "old" device might lack a serial, but since comparison only
looks at the cookie, that device won't be updated with a valid serial,
even if one is present in the new list.

So when creating the list, make sure that "unknown serial"s are updated.