fix: First letter of serialnr might be omitted
In some cases, where the serial number is provided to
devie-api-usb#manager, can have the first letter omitted. Since node-usb
will always include the full serial number, we are required to do a
contains check instead of equality to make sure we dont ignore targetd
deviced just because the first letter is ommitted.