fix: On reset received...continue read
The old implementation was done in such a way that when the device sent
a buffer that was less than the size of the header, we assumed a reset
sequence...tearing down the communication with the device. This seems to
have some impact especially during upgrade process where such scenario
can happen. After several testing, we came to conclusion that just
continuing the read when this scenario happens does not result in faulty
behavior.