fix: node-usb buffer split wont work on ubuntu
It seems that on ubuntu when we leave the buffer splitting resposibility
to node-usb default implementation, we get a LIBUSB_ERROR_IO back. This
happens during upload of HPK image to the device which in itself is a
large transfer that must be split in approx 1000 transfers. To
compensate for this, we are re-introducing out own chunk split
functionality but it will be invoked only when we receive a
LIBUSB_ERROR_IO and the data we are trying to send is larger then the
MAX_PACKET_SIZE. If libusb supports spliting large data transfers into
chunks then we proceed with that as default behavior.